+++ sas-20181203/ChangeLog.md
@@ -1,3 +1,22 @@
+2024-10-10 Version: 3.2.0
+- Support API GetBuildRiskDefineRuleConfig.
+- Support API SetBuildRiskDefineRuleConfig.
+- Update API DescribeExposedInstanceDetail: update response param.
+- Update API DescribeExposedInstanceList: add param AssetType.
+- Update API DescribeExposedInstanceList: add param CspmStatus.
+- Update API DescribeExposedInstanceList: update response param.
+- Update API DescribeExposedStatistics: update response param.
+- Update API DescribeOnceTask: add param Source.
+- Update API DescribeOnceTask: update response param.
+- Update API ListCheckResult: add param OperationTypes.
+- Update API ListCheckResult: update response param.
+- Update API ListSystemClientRules: update response param.
+- Update API SubmitCheck: update response param.
+- Update API SubmitOperationTask: add param RelationKey.
+- Update API SubmitOperationTask: update param DimensionType.
+- Update API VerifyCheckResult: update response param.
+
+
 2024-09-24 Version: 3.1.0
 - Support API AddPublishBatch.
 - Support API BatchCreateMaliciousNote.